tests/requirements.txt
# Test Requirements for Tourism Database
# =====================================

# Core testing framework
unittest2>=1.1.0

# Mocking and test utilities
mock>=4.0.3

# Database testing
psycopg2-binary>=2.8.6

# Web testing
flask>=2.0.0
flask-socketio>=5.0.0

# RDF/TTL parsing (for validation tests)
rdflib>=6.0.0

# Data manipulation and analysis
pandas>=1.3.0
numpy>=1.21.0

# HTTP testing
requests>=2.25.0
responses>=0.18.0

# Test coverage
coverage>=6.0

# Performance testing
pytest-benchmark>=3.4.1

# Parallel test execution
pytest-xdist>=2.4.0

# Test reporting
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# Memory profiling for performance tests
memory-profiler>=0.60.0

# Time utilities for test timing
freezegun>=1.2.0